refactor: isolate server entity
